// src/modules/calc/dto/update-calc.dto.ts
import { PartialType } from '@nestjs/mapped-types';
import { CreateCalcDto } from './create-calc.dto';

export class UpdateCalcDto extends PartialType(CreateCalcDto) {
  armazenamento_inicial_primeiro_dia?: number;
  total_dias_estagio?: number;
  armazenamento_final_dia_anterior?: number;
  lamina_liquida_previa_dia_anterior?: number;
  percentual_area_sombreada_previa?: number;
  percentual_area_sombreada_corrigida?: number;
  ponto_murcha_calculado?: number;
  capacidade_campo_calculado?: number;
  umidade_critica_percentual?: number;
  umidade_critica?: number;
  disponibilidade_total_agua?: number;
  lamina_armazenamento?: number;
  lamina_armazenada?: number;
  ks?: number;
  kc_ajustado?: number;
  kc_limite?: number;
  etc_dia_atual_cultura?: number;
  etc_dia_atual_cultura_corrigida?: number;
  lamina_liquida_sem_correcao?: number;
  lamina_liquida?: number;
  lamina_bruta?: number;
  intensidade_aplicacao?: number;
  tempo_irrigacao?: number;
  lamina_total?: number;
  agua_facilmente_disponivel?: number;
  armazenamento_inicial_dia_atual?: number;
  armazenamento_final_dia_atual?: number;
  excesso_deficit?: number;
  quantidade_total_agua?: number;
  volume_aplicado_area_total?: number;
  volume_aplicado_setor?: number;
  tempo_irrigacao_sugerido_area_total?: string;
  tempo_irrigacao_sugerido_area_setor?: string;
  created_at?: Date;
  primeiro_calculo?: number;
  decisao?: string;
  tipo_calculo?: string;
  calc_id?: number;
  id_cultura?: number;
}